+++ CudaSudoku.cu
@@ -4,8 +4,8 @@
 #include <cuda_runtime.h>
 #include "BitOperations.cu"
 
-__global__ void cudaBFSSudoku(uint64_t *old_boards,
-        uint64_t *new_boards,
+__global__ void cudaBFSSudoku(int *old_boards,
+        int *new_boards,
         int total_boards,
         int *board_index,
         int empty_row,
